ADD MahApps.Metro and MaterialDesignThemes.MahApps UPDATE MaterialDesignThemes to latest version
to get rid of the ugly white titlebar i added MahApps and updated every window xaml file with titlebar enabled to use the new MetroWindow from MahApps

in App.xaml i added a default window style named 'DalamudWindow' which every MetroWindow is using

i also changed DalamudInjectionDelayMs from decimal to double in ILauncherSettingsV3 for the new NumericUpDown control from MahApps replacing DecimalUpDown since the old one was ugly and unstyled

+++ src/XIVLauncher/Windows/SettingsControl.xaml.cs
@@ -88,7 +88,7 @@ namespace XIVLauncher.Windows
 
             VersionLabel.Text += " - v" + Util.GetAssemblyVersion() + " - " + Util.GetGitHash() + " - " + Environment.Version;
 
-            var val = (decimal) App.Settings.SpeedLimitBytes / BYTES_TO_MB;
+            var val = (double) App.Settings.SpeedLimitBytes / BYTES_TO_MB;
 
             SpeedLimiterUpDown.Value = val;
         }
